Made client abstract
