+++ src/Client.php
@@ -5,49 +5,14 @@ declare(strict_types=1);
 namespace Thesis\Grpc;
 
 use Amp\Cancellation;
-use Amp\Http\Client\DelegateHttpClient;
 use Amp\NullCancellation;
-use Thesis\Grpc\Client\Internal\Http2;
 use Thesis\Grpc\Exception\ClientStreamIsClosed;
-use Thesis\Protobuf\Decoder;
 
 /**
  * @api
  */
-final readonly class Client
+interface Client
 {
-    private Http2\InterceptorComposer $interceptor;
-
-    private Http2\StreamFactory $streams;
-
-    /**
-     * @param non-empty-string $host
-     * @param list<Client\Interceptor> $interceptors
-     */
-    public function __construct(
-        string $host,
-        DelegateHttpClient $client,
-        Encoding\Encoder $encoder,
-        Compression\Compressor $compressor,
-        Decoder $protobuf,
-        array $interceptors = [],
-    ) {
-        $this->interceptor = new Http2\InterceptorComposer([
-            ...$interceptors,
-            new Http2\AppendControlMetadataInterceptor(
-                $encoder->name(),
-                $compressor->name(),
-            ),
-        ]);
-        $this->streams = new Http2\StreamFactory(
-            http: $client,
-            uri: new Http2\UriFactory($host),
-            errors: new Http2\ErrorHandler($protobuf),
-            encoder: $encoder,
-            compressor: $compressor,
-        );
-    }
-
     /**
      * @template In of object
      * @template Out of object
@@ -62,18 +27,7 @@ final readonly class Client
         Client\Invoke $invoke,
         Metadata $md = new Metadata(),
         Cancellation $cancellation = new NullCancellation(),
-    ): object {
-        $stream = $this->createStream(
-            $invoke,
-            $md,
-            $cancellation,
-        );
-
-        $stream->send($request);
-        $stream->close();
-
-        return $stream->receive();
-    }
+    ): object;
 
     /**
      * @template In of object
@@ -85,13 +39,5 @@ final readonly class Client
         Client\Invoke $invoke,
         Metadata $md = new Metadata(),
         Cancellation $cancellation = new NullCancellation(),
-    ): ClientStream {
-        /** @var ClientStream<In, Out> */
-        return $this->interceptor->intercept(
-            $invoke,
-            $md,
-            $cancellation,
-            $this->streams->create(...),
-        );
-    }
+    ): ClientStream;
 }
